+++ Transformation/Transformation/Vector3.h
@@ -7,9 +7,9 @@ class Vector3
 {
 public: //attributes 
 
-	double x{ 0.0 };
-	double y{ 0.0 };
-	double z{ 0.0 };
+	float x{ 0.0 };
+	float y{ 0.0 };
+	float z{ 0.0 };
 
 
 public: //functions
@@ -19,7 +19,7 @@ public: //functions
 	Vector3() = default;
 
 	// constructor with 3 values 
-	Vector3(double x1, double y1, double z1);
+	Vector3(float x1, float y1, float z1);
 
 	// copy connstructor
 	Vector3(Vector3 const&);
@@ -29,9 +29,9 @@ public: //functions
 
 	//OTHER
 
-	double length() const;
+	float length() const;
 
-	double lenghtSquared() const;
+	float lenghtSquared() const;
 
 	//Reduce the length to 1
 	void Normalise();
@@ -52,7 +52,7 @@ public: //functions
 	friend Vector3 operator*(Vector3 const& v1, Vector3 const& v2);
 
 	//product of a scalar by a vector
-	friend Vector3 operator*(double k, Vector3 const& v1);
+	friend Vector3 operator*(float k, Vector3 const& v1);
 
 	//Product of a scalat by a vector
 	friend Vector3 operator*(int k, Vector3 const& v1);
